refactor in preparation for moving cast to spark-expr crate
